requirements.txt
uvicorn==0.34.3
fastapi==0.115.12
python-dotenv==1.1.0
pydub==0.25.1
librosa==0.11.0
zhon==2.1.1
loguru==0.7.3
ruamel.yaml==0.18.13
HyperPyYAML==1.2.2
modelscope==1.26.0
omegaconf==2.3.0
hydra-core==1.3.2
gdown==5.2.0
matplotlib==3.10.3
wget==3.2
pyarrow==20.0.0
pyworld==0.3.5
gradio==5.33.0
pyloudnorm==0.1.1
rich==14.1.0
seg2stream==1.0.7
##########
onnxruntime-gpu==1.20.2
openai-whisper==20240930
wetext==0.0.8
lightning==2.5.1.post0
conformer==0.3.2
diffusers==0.33.1
tensorrt-cu12==10.0.1
tensorrt-cu12-bindings==10.0.1
tensorrt-cu12-libs==10.0.1
peft==0.15.2
##########
numpy==1.26.4
ray==2.40.0
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.5.1
vllm==0.7.3
flashinfer-python==0.2.2
# flashinfer-python==0.2.2 --index-url https://flashinfer.ai/whl/cu124/torch2.5/
